Reorganize files layout
