+++ src/mod.ts
@@ -1,6 +1,7 @@
 /**
  * @module @run-z/ndk
  */
+export * from './fs/mod.js';
 export * from './package-entry-point.js';
 export * from './package-entry-targets.js';
 export * from './package-info.js';
