+++ src/resolution/mod.ts
@@ -1,9 +1,6 @@
 export * from './dependency-resolution.js';
 export * from './import-resolution.js';
 export * from './import.js';
-export * from './node-package-fs.js';
-export * from './package-fs.js';
 export * from './package-resolution.js';
 export * from './recognize-import.js';
 export * from './sub-package-resolution.js';
-export * from './virtual-package-fs.js';
