+++ src/resolution/package-resolution.spec.ts
@@ -1,7 +1,8 @@
 import { beforeEach, describe, expect, it } from '@jest/globals';
+import { resolveRootPackage } from '../fs/resolve-root-package.js';
+import { VirtualPackageFS } from '../fs/virtual-package-fs.js';
 import { PackageInfo } from '../package-info.js';
-import { PackageResolution, resolveRootPackage } from './package-resolution.js';
-import { VirtualPackageFS } from './virtual-package-fs.js';
+import { PackageResolution } from './package-resolution.js';
 
 describe('PackageResolution', () => {
   let fs: VirtualPackageFS;
